Revert "Working simple slider, related to issue #378"
