--- src/components/bubbleopacity/_simpleslider.scss
@@ -1,84 +1,0 @@
-//general styles
-.vzb-ss-holder {
-
-  height: 20px;
-  margin: 0 auto;
-  position: relative;
-
-  &:before {
-    @include border-radius(5px 5px 5px 5px);
-    content: '';
-    position: absolute;
-    top: 50%;
-    margin-top: -4px;
-    width: 100%;
-    height: 6px;
-    background-color: rgba($vzb-color-primary, .3);
-  }
-
-  &.vzb-disabled {
-    &:before {
-      background-color: rgba($vzb-color-grey, .2);
-    }
-    .vzb-ss-slider {
-      opacity: .5;
-    }
-  }
-
-  .vzb-ss-slider {
-    position: absolute;
-    top: 50%;
-    margin-top: -3px;
-    height: 3px;
-    width: 100%;
-    -webkit-appearance: none;
-    background: transparent;
-    outline: none !important;
-    border: 0;
-
-    // Slider Handle
-    &::-webkit-slider-thumb {
-      @include border-radius(20px);
-      @include transition(left .1s ease-in-out);
-      -webkit-appearance: none;
-      cursor: pointer;
-      background-color: $vzb-color-primary;
-      top: -12px;
-      width: 20px;
-      height: 20px;
-      border: 3px solid $vzb-color-white;
-    }
-    &::-moz-range-thumb {
-      @include border-radius(20px);
-      @include transition(left .1s ease-in-out);
-      cursor: pointer;
-      background: $vzb-color-primary;
-      top: -12px;
-      width: 20px;
-      height: 20px;
-      border: 3px solid $vzb-color-white;
-    }
-
-  }
-}
-
-//large specific styles
-&.vzb-large {
-  .vzb-ss-holder {
-    //...
-  }
-}
-
-//medium specific styles
-&.vzb-medium {
-  .vzb-ss-holder {
-    //...
-  }
-}
-
-//small specific styles
-&.vzb-small {
-  .vzb-ss-holder {
-    //...
-  }
-}
