+++ src/components/buttonlist/_buttonlist.scss
@@ -12,7 +12,7 @@ $vzb-dialog-content-height: 250px;
 .vzb-buttonlist {
   @import '../bubblesize/bubblesize';
   @import '../indicatorpicker/indicatorpicker';
-  @import '../simpleslider/simpleslider';
+  @import '../bubbleopacity/bubbleopacity';
   @import '../colorlegend/colorlegend';
   @import '../simplecheckbox/simplecheckbox';
   @import '../draggablelist/draggablelist';
