+++ src/components/buttonlist/dialogs/moreoptions/moreoptions.html
@@ -21,8 +21,8 @@
     <div class="vzb-yaxis-container"></div>
 
     <p class="vzb-dialog-sublabel"><%=t ("buttons/size") %></p>
-    <div class="vzb-saxis-container"></div>
-
+    <div class="vzb-saxis-container"></div> 
+      
     <div class="vzb-dialog-br"></div>
 
     <p class="vzb-dialog-sublabel"><%=t ("buttons/presentation") %></p>
