+++ src/components/timeslider/timeslider.html
@@ -9,7 +9,15 @@
         </g>
       </g>
       </svg>
-      <div class="vzb-ts-delay-slider"></div>
+      <svg class="vzb-ts-delay-slider">
+        <g>
+          <g class="vzb-ts-delay-slider-axis"></g>
+          <g class="vzb-ts-delay-slider-slide">
+            <circle class="vzb-ts-delay-slider-handle"></circle>
+            <text class="vzb-ts-delay-slider-value"></text>
+          </g>
+        </g>
+        </svg>
   </div>
 
   <!-- SVG VIZABI ICONS
